Reorganize backend to subsystem-first architecture
- Move domain-specific files from shared/ to appropriate subsystems:
  - Banking: enums.js, ManualCategorized.js, Tag.js, dataSyncService.js, enhanced-keyword-matching/, tagService.js, installmentGroupingUtils.js
  - Investments: investmentConstants.js
  - Monthly-budgets: patternTypes.js, Category.js, SubCategory.js, userCategoryService.js, patternMatching.js
  - Project-budgets: statusTypes.js
  - Translation: Translation.js, translationService.js
- Update import paths across all affected files
- Keep truly cross-cutting concerns in shared/ (config, middleware, utils)
