+++ backend/src/banking/models/ManualCategorized.js
