+++ backend/src/banking/utils/installmentGroupingUtils.js
