+++ backend/src/investments/constants/investmentConstants.js
