+++ backend/src/investments/routes/investments.js
@@ -3,7 +3,7 @@ const router = express.Router();
 const investmentService = require('../services/investmentService');
 const bankScraperService = require('../../banking/services/bankScraperService');
 const startupInvestmentService = require('../services/startupInvestmentService');
-const dataSyncService = require('../../shared/services/dataSyncService');
+const dataSyncService = require('../../banking/services/dataSyncService');
 const { BankAccount } = require('../../shared/models');
 const auth = require('../../shared/middleware/auth');
 const logger = require('../../shared/utils/logger');
