+++ backend/src/monthly-budgets/services/budgetCalculationService.js
@@ -3,7 +3,7 @@ const TransactionPattern = require('../models/TransactionPattern');
 const logger = require('../../shared/utils/logger');
 const averagingDenominatorService = require('./averagingDenominatorService');
 const PatternMatchingUtils = require('../../shared/utils/patternMatching');
-const { PATTERN_TYPES } = require('../../shared/constants/patternTypes');
+const { PATTERN_TYPES } = require('../constants/patternTypes');
 
 class BudgetCalculationService {
   /**
