+++ backend/src/monthly-budgets/services/recurrenceDetectionService.js
@@ -2,7 +2,7 @@ const { Transaction } = require('../../shared/models');
 const TransactionPattern = require('../models/TransactionPattern');
 const logger = require('../../shared/utils/logger');
 const { v4: uuidv4 } = require('uuid');
-const { PATTERN_TYPES } = require('../../shared/constants/patternTypes');
+const { PATTERN_TYPES } = require('../constants/patternTypes');
 
 class RecurrenceDetectionService {
   /**
